# depot-lifecycle-openapi-2.2.5.yaml
openapi: 3.0.1
info:
  title: Depot Life Cycle
  description: "# Purpose\n\n A depot centric API for managing the interchange activity\
    \ & repair lifecycle of a shipping container.  The API is expected to be used\
    \ by Customers, Depots, and Owners to facilitate real time communication between\
    \ systems instead of traditional EDI files.\n\n\n\n # Overview\n\n The depot lifecycle\
    \ API is a RESTful API.  It's requests & responses are loosely based upon traditional\
    \ EDI files.  For code definitions, explanations, or traditional EDI definitions\
    \ refer to [IICL TB 002, February 2003](https://www.iicl.org/iiclforms/assets/File/public/bulletins/TB002_EDIS_February_2003.pdf).\n\
    \n\n\n # JSON\n\n The requests and responses of this API are formatted according\
    \ to the [JSON](https://www.json.org/) standard.\n\n ## Null vs Absent\n\n For\
    \ all types other than arrays, nullable and absent are equivalent by this specification.\
    \  For example, if a non-array attribute is marked optional it may be excluded\
    \ from the JSON or it may be explicitly set to null.  However, to keep request\
    \ sizes small, it is recommended to prefer excluding attributes rather than sending\
    \ an explicit null.\n\n For arrays, null should be avoided.  For example, if the\
    \ array is empty or there are no values expected for the array then either the\
    \ field can be excluded from the JSON or an empty array should be transmitted.\n\
...
  contact:
...
servers:
...
security:
- Dynamic_Token: []
- Static_Token: []
...
paths:
...
  /api/v2/gate/{unitNumber}:
    get:
      tags:
      - gate
      summary: fetch the current gate status
      description: "For the given unit number, if the shipping container is currently\
        \ gated in or gated out, fetch the current interchange information - status,\
        \ the time of the gate, etc.  This is a reporting only function.  To correct\
        \ a gate status, use the update a gate record endpoint."
      operationId: showGate
      parameters:
      - name: unitNumber
        in: path
        description: the current unit number of the shipping container
        required: true
        schema:
          maxLength: 11
          pattern: "^[A-Z]{4}[X0-9]{6}[A-Z0-9]{0,1}$"
          type: string
          example: CONU1234561
      responses:
        "200":
          description: successfully found current gate status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GateStatus"
        "400":
          description: an error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: fetching a gate record is disallowed by security
        "404":
          description: the shipping container was not found
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
      x-iicl-purpose: reporting
  /api/v2/gatePhoto/{relatedId}:
    post:
      tags:
      - gate proposals
      summary: upload a gate photo
      description: "Instead of using a link, upload a photo for a previously created\
        \ gate record"
      operationId: uploadGatePhoto
      parameters:
      - name: relatedId
        in: path
        description: the related identifier (from the gate record) that this photo
          should be attached
        required: true
        schema:
          type: integer
          format: int64
          example: 10102561
      requestBody:
        description: The photo to upload (expected name of part is `file`)
        content:
          multipart/form-data:
            schema:
              type: string
              properties:
                file:
                  type: string
                  format: binary
              description: the photo data
              format: binary
            encoding:
              file:
                contentType: application/octet-stream
        required: true
      responses:
        "200":
          description: successfully uploaded a gate photo
          content:
            application/json:
              schema: {}
        "202":
          description: "photo accepted for processing, but not created due to manual\
            \ processing requirement"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingResponse"
        "400":
          description: an invalid request was provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: uploading a gate photo is disallowed by security
        "404":
          description: the gate record could not be found
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
      x-iicl-purpose: activity
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          maxLength: 6
          pattern: "^[A-Z0-9]{3}[0-9]{3}$"
          required:
          - "false"
          type: string
          description: indicator code specific to this error
          nullable: true
          example: TRI521
        message:
          required:
          - "false"
          type: string
          description: a descriptive error message
          nullable: true
          example: Info TRI521 - Unit has been gated-in but is not off-hired
        details:
          required:
          - "false"
          type: array
          description: details for the descriptive error message
          nullable: false
          items:
            type: string
            example: "Customer may turn in unit, but will continued to be billed."
      description: |-
        a general response object used when an error occurs to describe why it occurred

        `400` http status codes are used when a request is invalid or the basic requirements are not met

        `405` http status codes are used for business validations
...
    GateStatus:
      required:
      - activityTime
      - adviceNumber
      - depot
      - status
      type: object
      properties:
        relatedId:
          required:
          - "false"
          type: integer
          description: |-
            *Field is currently proposed to be added - not currently production approved.*

            An internal system identifier to be used to upload Gate Photos or compare related activities.
          format: int64
          nullable: true
          example: 10102561
        adviceNumber:
          maxLength: 16
          required:
          - "true"
          type: string
          description: the redelivery or release advice number for the gate record
          nullable: false
          example: AHAMG000000
        depot:
          description: the storage location for the given advice number
          allOf:
          - $ref: "#/components/schemas/Party"
        status:
          maxLength: 1
          required:
          - "true"
          type: string
          description: |-
            the current status

            `A` - Non-damaged

            `D` - Damaged

            `S` - Sold
          nullable: false
          example: D
          enum:
          - A
          - D
          - S
        activityTime:
          required:
          - "true"
          type: string
          description: |-
            the date and time of the gate activity in local time

            ( notation as defined by [RFC 3339, section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6) )
          format: date-time
          nullable: false
        insuranceCoverage:
          description: The applicable insurance coverage for damage estimate purposes
          nullable: true
          allOf:
          - $ref: "#/components/schemas/InsuranceCoverage"
        currentExchangeRate:
          required:
          - "false"
          type: number
          description: the exchange rate to convert billed currency to the local currency
            for damage estimate totals
          format: double
          nullable: true
          example: 0.8133
        currentInspectionCriteria:
          maxLength: 10
          required:
          - "false"
          type: string
          description: "the unit should be inspected to this grade; if none, no estimate\
            \ is allowed"
          nullable: true
          example: IICL
        type:
          maxLength: 3
          required:
          - "false"
          type: string
          description: |-
            current gate type indicator

            `IN` - Gate In

            `OUT` - Gate Out

            No value means unknown.
          nullable: true
          example: IN
          enum:
          - IN
          - OUT
      description: the current gate status of a given shipping container
...
    InsuranceCoverage:
      type: object
      properties:
        amountCovered:
          required:
          - "false"
          type: number
          description: the total amount allowable to by covered by insurance
          format: double
          nullable: true
        amountCurrency:
          pattern: "^[A-Z]{3}$"
          required:
          - "false"
          type: string
          description: The currency for the insurance coverage amount
          nullable: true
          example: EUR
        appliesToCTL:
          required:
          - "false"
          type: boolean
          description: indicates if this coverage should apply to scenarios where
            the damage exceeds the depreciated value of the container
          nullable: true
        allOrNothing:
          required:
          - "false"
          type: boolean
          description: "indicates if the amount of damage exceeds the insurance coverage,\
            \ whether the lessee if fully responsible for the damage"
          nullable: true
        exceptions:
          required:
          - "false"
          type: array
          description: any contractual reason why the coverage may not apply
          nullable: false
          items:
            type: string
        exclusions:
          required:
          - "false"
          type: array
          description: reasons insurance coverage would be excluded from a repair
          nullable: false
          items:
            type: string
        inclusions:
          required:
          - "false"
          type: array
          description: reasons insurance coverage would include a repair
          nullable: false
          items:
            type: string
      description: describes the type of insurance coverage a given shipping container
        has for a damage repair
...
    Party:
      required:
      - companyId
      type: object
      properties:
        companyId:
          maxLength: 9
          pattern: "^[A-Z0-9]{9}$"
          required:
          - "true"
          type: string
          description: "the identifier for this party, often referred to as an EDI\
            \ Address"
          nullable: false
          example: DEHAMCMRA
        userCode:
          maxLength: 16
          required:
          - "false"
          type: string
          description: the user identifier at this company that concerns this message
          nullable: true
          example: JDOE
        userName:
          maxLength: 70
          required:
          - "false"
          type: string
          description: the full name for the user identified by `userCode`
          nullable: true
          example: John Doe
        faxNumber:
          required:
          - "false"
          type: array
          description: the contact fax number(s) for this party
          nullable: false
          items:
            type: string
        phoneNumber:
          required:
          - "false"
          type: array
          description: the contact phone number(s) for this party
          nullable: false
          items:
            type: string
        emailAddress:
          required:
          - "false"
          type: array
          description: the contact email address(es) for this party
          nullable: false
          items:
            type: string
        name:
          maxLength: 150
          required:
          - "false"
          type: string
          description: the name of this company
          nullable: true
          example: CMR Container Maintenance Rep.
        code:
          maxLength: 10
          required:
          - "false"
          type: string
          description: "the internal system code for this company, will be system\
            \ specific to the system delivering or receiving this message"
          nullable: true
          example: HAMG
        streetAddress:
          required:
          - "false"
          type: array
          description: the street address lines
          nullable: false
          items:
            type: string
        city:
          maxLength: 28
          required:
          - "false"
          type: string
          description: the city for the address
          nullable: true
        country:
          maxLength: 2
          required:
          - "false"
          type: string
          description: the 2 digit ISO country code for the address
          nullable: true
        postalCode:
          maxLength: 20
          required:
          - "false"
          type: string
          description: the postal code for the address
          nullable: true
        stateProvince:
          maxLength: 20
          required:
          - "false"
          type: string
          description: the optional state or province code for the address
          nullable: true
        latitude:
          required:
          - "false"
          type: number
          description: "Instead of an address, provides the latitude of this party"
          format: double
          nullable: true
        longitude:
          required:
          - "false"
          type: number
          description: "Instead of an address, provides the longitude of this party"
          format: double
          nullable: true
      description: represents a company (or location) involving shipping containers
    PendingResponse:
      type: object
      properties:
        code:
          maxLength: 6
          pattern: "^[A-Z0-9]{3}[0-9]{3}$"
          required:
          - "false"
          type: string
          description: indicator code for this response
          nullable: true
          example: EXX365
        message:
          required:
          - "false"
          type: string
          description: a descriptive reason why the activity was not immediately created
          nullable: true
          example: "Info EXX365 - Accepted, but requires verification to process."
      description: used inform why an activity could not be immediately processed
...
  securitySchemes:
    JWT_TOKEN:
      type: http
      description: JWT Bearer Authentication
      scheme: bearer
      bearerFormat: JWT
...